Changes in sensor interfaces. Now biases are float since used only for cooked data.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@9800 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/lib/peripherals/sensors/hal_barometer.h
@@ -48,7 +48,7 @@
  */
 #define _base_barometer_methods_alone                                       \
   /* Invoke the set bias procedure.*/                                       \
-  msg_t (*set_bias)(void *instance, int32_t biases[]);                      \
+  msg_t (*set_bias)(void *instance, float biases[]);                        \
   /* Remove bias stored data.*/                                             \
   msg_t (*reset_bias)(void *instance);                                      \
   /* Invoke the set sensitivity procedure.*/                                \
